requirements.txt
django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
drf_yasg